Move single-package utilities out of @internals/utils
@internals/utils held helpers used by exactly one downstream package.
Moved each into its sole consumer, colocated with its tests:

- @kubb/cli: canUseTTY/isCIEnvironment, the mascot banner and duration
  color helpers, toCause, and createSerialRunner
- @kubb/core: BuildError, isPathInside, toFilePath, inParallel, and
  randomCliColor
- @kubb/kit: the Url helper and its reserved-word dependency

Left matchesStored and isToolAvailable in @internals/utils: both back a
shared helper (write, detectTool) used across packages, even though only
one package also imports them directly.
