+++ internals/utils/src/index.ts
@@ -1,12 +1,9 @@
 export { camelCase, pascalCase } from './casing.ts'
-export { formatMsWithColor, getIntro, randomCliColor } from './colors.ts'
-export { canUseTTY, isCIEnvironment } from './env.ts'
-export { BuildError, getErrorMessage, toCause, toError } from './errors.ts'
-export { clean, exists, isPathInside, matchesStored, read, toFilePath, toPosixPath, trimExtName, write } from './fs.ts'
+export { getErrorMessage, toError } from './errors.ts'
+export { clean, exists, matchesStored, read, toPosixPath, trimExtName, write } from './fs.ts'
 export type { PossiblePromise } from './promise.ts'
-export { createSerialRunner, inParallel, isPromise, memoize } from './promise.ts'
+export { isPromise, memoize } from './promise.ts'
 export type { LiteralUnion } from './types.ts'
 export { runtime } from './runtime.ts'
 export { detectTool, formatters, isToolAvailable, linters, tokenize, type ToolCommand, FORMATTER_PREFERENCE, LINTER_PREFERENCE } from './tools.ts'
 export { formatMs, getElapsedMs } from './time.ts'
-export { Url } from './Url.ts'
