+++ internals/utils/src/promise.ts
@@ -1,5 +1,3 @@
-import { toError } from './errors.ts'
-
 /** A value that may already be resolved or still pending.
  *
  * @example
@@ -68,84 +66,3 @@ export function memoize<TKey, TValue>(store: Store<TKey, TValue>, factory: (key:
   }
 }
 
-type SerialRunnerOptions = {
-  /**
-   * The async work to serialize.
-   */
-  run(): Promise<void>
-  /**
-   * Receives errors thrown by `run`, so a failure never rejects the returned trigger.
-   */
-  onError(error: Error): void
-}
-
-/**
- * Wraps `run` so invocations never overlap: a trigger that lands while a run is in flight
- * marks it dirty and runs once more after it finishes, no matter how many triggers arrived.
- * Useful for event-driven reruns (a file watcher, a queue drain) where bursts should
- * coalesce into a single trailing run.
- *
- * @example
- * ```ts
- * const rebuild = createSerialRunner({
- *   run: () => build(),
- *   onError: (error) => log.error(error.message),
- * })
- * watcher.on('change', () => void rebuild())
- * ```
- */
-export function createSerialRunner({ run, onError }: SerialRunnerOptions): () => Promise<void> {
-  let running = false
-  let dirty = false
-
-  return async (): Promise<void> => {
-    if (running) {
-      dirty = true
-      return
-    }
-    running = true
-    do {
-      dirty = false
-      try {
-        await run()
-      } catch (error) {
-        onError(toError(error))
-      }
-    } while (dirty)
-    running = false
-  }
-}
-
-type ParallelOptions<TItem> = {
-  /**
-   * Items to work through, handed out in order.
-   */
-  items: ReadonlyArray<TItem>
-  /**
-   * How many items may be in flight at once.
-   */
-  limit: number
-  /**
-   * Runs once per item, with the item's position so a caller can report progress.
-   */
-  run(item: TItem, index: number): Promise<void>
-}
-
-/**
- * Runs `run` over every item with at most `limit` in flight. Workers share one iterator, so each
- * takes the next item the moment it frees up instead of waiting for a batch to drain.
- *
- * @example
- * ```ts
- * await inParallel({ items: files, limit: 50, run: (file) => storage.writeItem(file.path, file.source) })
- * ```
- */
-export async function inParallel<TItem>({ items, limit, run }: ParallelOptions<TItem>): Promise<void> {
-  const queue = items.entries()
-
-  const worker = async (): Promise<void> => {
-    for (const [index, item] of queue) await run(item, index)
-  }
-
-  await Promise.all(Array.from({ length: Math.min(limit, items.length) }, () => worker()))
-}
