+++ packages/cli/src/loggers/utils.ts
@@ -1,8 +1,9 @@
 import process from 'node:process'
 import { styleText } from 'node:util'
-import { canUseTTY, formatMs, getElapsedMs, toCause } from '@internals/utils'
+import { formatMs, getElapsedMs } from '@internals/utils'
 import type { Config, Reporter, ReporterContext } from '@kubb/core'
 import { logLevel as logLevelMap } from '@kubb/core'
+import { canUseTTY } from '../utils/env.ts'
 import type { LoggerContext, LoggerOptions } from './defineLogger.ts'
 import { clackLogger } from './clackLogger.ts'
 import { plainLogger } from './plainLogger.ts'
@@ -32,6 +33,13 @@ export function formatMessage(message: string, logLevel: number): string {
     return `${styleText('dim', `[${timestamp}]`)} ${message}`
   }
   return message
+}
+
+/**
+ * Extracts the `.cause` of an `Error` as an `Error`, or `undefined` when absent or not an `Error`.
+ */
+function toCause(error: Error): Error | undefined {
+  return error.cause instanceof Error ? error.cause : undefined
 }
 
 /**
