+++ packages/cli/src/runners/studio/run.ts
@@ -4,7 +4,7 @@ import process from 'node:process'
 import { styleText } from 'node:util'
 import * as prompts from '@clack/prompts'
 import { KUBB_CONFIG_FILENAME } from '@internals/shared'
-import { canUseTTY, isCIEnvironment, toError } from '@internals/utils'
+import { toError } from '@internals/utils'
 import type { CLIOptions, Config } from '@kubb/core'
 import { createFileStorage, createClient, defaultStudioUrl, InvalidAgentTokenError, pollForPairingToken, setStorage, startPairing } from '@kubb/studio'
 import { x } from 'tinyexec'
@@ -12,6 +12,7 @@ import type { CommandRunner } from 'gunshi'
 import { buildTelemetryEvent, sendTelemetry } from '../../Telemetry.ts'
 import { version } from '../../../package.json'
 import type { definition } from '../../commands/studio.ts'
+import { canUseTTY, isCIEnvironment } from '../../utils/env.ts'
 import { getConfigs } from '../generate/utils.ts'
 import { clearCredentials, type Credentials, getKubbHome, readCredentials, writeCredentials } from './credentials.ts'
