+++ packages/cli/src/utils/env.test.ts
